LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/remove-from-document-expected.txt

CONSOLE MESSAGE: line 29: TypeError: null is not an object (evaluating 'sourceIFrame.contentDocument.querySelector("meta").remove')

Harness Error (FAIL), message = TypeError: null is not an object (evaluating 'sourceIFrame.contentDocument.querySelector("meta").remove')
